refactor: extract routers/admin.py
Move /admin/* routes out of main.py into api/routers/admin.py. Extract
_create_github_issue helper so the escalate handler stays thin (~10 lines).
Wire router into main.py.
